requirements.txt
streamlit>=1.34.0
pandas>=2.0.0
requests>=2.31.0
pymupdf>=1.20.0
pytesseract>=0.3.10
pdf2image>=1.16.3
pillow>=10.0.0
langchain-openai>=0.1.6
python-dotenv>=1.0.0
openpyxl>=3.0.10
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0
